Start refactoring the check unit tests into more manageable class
